Made the box render into a component
